# HBASE-15519 Add per-user metrics with lossy counting