And extra commit   "HBASE-23054 Remove synchronization block from MetaTableMetrics and fix LossyCounting algorithm" is just only to support the changes done in HBASE-15519 and is separately getting reviewed on HBASE-23054 Jira itself.