@Apache-HBase 2019-09-25
:broken_heart: **-1 overall**






| Vote | Subsystem | Runtime | Comment |
|:----:|----------:|--------:|:--------|
| :blue_heart: |  reexec  |   0m 30s |  Docker mode activated.  |
||| _ Prechecks _ |
| :green_heart: |  dupname  |   0m  0s |  No case conflicting files found.  |
| :green_heart: |  hbaseanti  |   0m  0s |  Patch does not have any anti-patterns.  |
| :green_heart: |  @author  |   0m  0s |  The patch does not contain any @author tags.  |
| :green_heart: |  test4tests  |   0m  0s |  The patch appears to include 4 new or modified test files.  |
||| _ master Compile Tests _ |
| :blue_heart: |  mvndep  |   1m  0s |  Maven dependency ordering for branch  |
| :green_heart: |  mvninstall  |   5m 41s |  master passed  |
| :green_heart: |  compile  |   1m 28s |  master passed  |
| :green_heart: |  checkstyle  |   1m 50s |  master passed  |
| :green_heart: |  shadedjars  |   5m 16s |  branch has no errors when building our shaded downstream artifacts.  |
| :green_heart: |  javadoc  |   1m 15s |  master passed  |
| :blue_heart: |  spotbugs  |   4m 42s |  Used deprecated FindBugs config; considering switching to SpotBugs.  |
| :green_heart: |  findbugs  |   5m 47s |  master passed  |
||| _ Patch Compile Tests _ |
| :blue_heart: |  mvndep  |   0m 16s |  Maven dependency ordering for patch  |
| :green_heart: |  mvninstall  |   5m 52s |  the patch passed  |
| :green_heart: |  compile  |   1m 40s |  the patch passed  |
| :green_heart: |  javac  |   1m 40s |  the patch passed  |
| :broken_heart: |  checkstyle  |   0m 12s |  hbase-hadoop-compat: The patch generated 5 new + 0 unchanged - 0 fixed = 5 total (was 0)  |
| :broken_heart: |  checkstyle  |   0m 13s |  hbase-hadoop2-compat: The patch generated 8 new + 0 unchanged - 0 fixed = 8 total (was 0)  |
| :broken_heart: |  checkstyle  |   1m 31s |  hbase-server: The patch generated 3 new + 2 unchanged - 0 fixed = 5 total (was 2)  |
| :green_heart: |  whitespace  |   0m  0s |  The patch has no whitespace issues.  |
| :green_heart: |  shadedjars  |   5m  5s |  patch has no errors when building our shaded downstream artifacts.  |
| :green_heart: |  hadoopcheck  |  17m  2s |  Patch does not cause any errors with Hadoop 2.8.5 2.9.2 or 3.1.2.  |
| :green_heart: |  javadoc  |   1m  4s |  the patch passed  |
| :green_heart: |  findbugs  |   5m 44s |  the patch passed  |
||| _ Other Tests _ |
| :green_heart: |  unit  |   0m 30s |  hbase-hadoop-compat in the patch passed.  |
| :broken_heart: |  unit  |   0m 33s |  hbase-hadoop2-compat in the patch failed.  |
| :broken_heart: |  unit  |  29m 29s |  hbase-server in the patch failed.  |
| :green_heart: |  asflicense  |   0m 40s |  The patch does not generate ASF License warnings.  |
|  |   | 100m  8s |   |


| Reason | Tests |
|-------:|:------|
| Failed junit tests | hadoop.hbase.regionserver.TestMetricsUserSourceImpl |
|   | hadoop.hbase.regionserver.TestMetricsUserAggregate |


| Subsystem | Report/Notes |
|----------:|:-------------|
| Docker | Client=19.03.1 Server=19.03.1 base: https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/artifact/out/Dockerfile |
| GITHUB PR | https://github.com/apache/hbase/pull/661 |
| Optional Tests | dupname asflicense javac javadoc unit spotbugs findbugs shadedjars hadoopcheck hbaseanti checkstyle compile |
| uname | Linux 13a1a50cf9cd 4.15.0-60-generic #67-Ubuntu SMP Thu Aug 22 16:55:30 UTC 2019 x86_64 GNU/Linux |
| Build tool | maven |
| Personality | /home/jenkins/jenkins-slave/workspace/HBase-PreCommit-GitHub-PR_PR-661/out/precommit/personality/provided.sh |
| git revision | master / 52f5a85bfc |
| Default Java | 1.8.0_181 |
| checkstyle | https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/artifact/out/diff-checkstyle-hbase-hadoop-compat.txt |
| checkstyle | https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/artifact/out/diff-checkstyle-hbase-hadoop2-compat.txt |
| checkstyle | https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/artifact/out/diff-checkstyle-hbase-server.txt |
| unit | https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/artifact/out/patch-unit-hbase-hadoop2-compat.txt |
| unit | https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/artifact/out/patch-unit-hbase-server.txt |
|  Test Results | https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/testReport/ |
| Max. process+thread count | 641 (vs. ulimit of 10000) |
| modules | C: hbase-hadoop-compat hbase-hadoop2-compat hbase-server U: . |
| Console output | https://builds.apache.org/job/HBase-PreCommit-GitHub-PR/job/PR-661/1/console |
| versions | git=2.11.0 maven=2018-06-17T18:33:14Z) findbugs=3.1.11 |
| Powered by | Apache Yetus 0.11.0 https://yetus.apache.org |


This message was automatically generated.